% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeBracket.R
\name{agebracket}
\alias{agebracket}
\title{agebracket}
\usage{
agebracket(age)
}
\arguments{
\item{age}{input value age}
}
\value{
the appropriate age bracket that the input age belongs to
}
\description{
agebracket
}
\examples{
\dontrun{
agebracket(42)
}

}
