% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeBracketCol.R
\name{agebracketcol}
\alias{agebracketcol}
\title{agebracketcol}
\usage{
agebracketcol(dataset)
}
\arguments{
\item{dataset}{the dataset on which an agebracket column will be added}
}
\value{
the treated dataset
}
\description{
agebracketcol
}
\examples{
\dontrun{
agebracketcol(data_all)
}
}
