% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeFilter.R
\name{agefilter}
\alias{agefilter}
\title{agefilter}
\usage{
agefilter(age, dataset)
}
\arguments{
\item{age}{input value age}

\item{dataset}{dataset to be treated}
}
\value{
a treated dataset after the appropriate age brackets have been filtered out
}
\description{
agefilter
}
\examples{
\dontrun{
agefilter(34, data_all)
}
}
