% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom20percent.R
\name{bottom20percent}
\alias{bottom20percent}
\title{bottom20percent}
\usage{
bottom20percent(dataset)
}
\arguments{
\item{dataset}{to be treated, the bottom 20 percent of the runners in terms of their finishing time will be returned}
}
\value{
the bottom 20 percent of the runners in terms of their finishing time
}
\description{
bottom20percent
}
\examples{
\dontrun{
bottom20percent(data_all)
}
}
