% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom20percentmean.R
\name{bottom20percentmean}
\alias{bottom20percentmean}
\title{bottom20percentmean}
\usage{
bottom20percentmean(dataset)
}
\arguments{
\item{dataset}{the dataset to be treated, the mean time at the specified milestones for the bottom 20 percent runners will be calcuated}
}
\value{
the treated dataset in the format of a dataframe
}
\description{
bottom20percentmean
}
\examples{
\dontrun{
bottom20percentmean(data_all)
}
}
