% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics_filter.R
\name{demographics_filter}
\alias{demographics_filter}
\title{demographics_filter}
\usage{
demographics_filter(dataset, age, gender, nationality)
}
\arguments{
\item{dataset}{the dataset to be treated, compulsory argument, cannot be empty}

\item{age}{input value age, the data will be filtered according to the appropriate age group of the input value. Optional argument}

\item{gender}{input value = "M" or "F", the data will be filtered accordingly. Optional argument}

\item{nationality}{input value nationality, the data will be filtered accordingly. Optional argument}
}
\value{
filtered dataset with only desired data points remaining
}
\description{
demographics_filter
}
\examples{
\dontrun{
demographics_filter(data_all, age = 34, gender = "F", nationality ="USA")
demographics_filter(data_all, gender = "F")
}
}
