% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalaverage.R
\name{globalaverage}
\alias{globalaverage}
\title{globalaverage}
\usage{
globalaverage(dataset)
}
\arguments{
\item{dataset}{the data set to be treated, the global mean time at the specified milestones will be calculated for all runners in the data set}
}
\value{
the mean time at the specified milestones for all the runners in the data set
}
\description{
globalaverage
}
\examples{
\dontrun{
globalaverage(data_all)
}
}
