% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goaltime.R
\name{goaltime}
\alias{goaltime}
\title{goaltime}
\usage{
goaltime(goaltime, dataset)
}
\arguments{
\item{goaltime}{your projected time spent to reach the finish line}

\item{dataset}{the dataset from which information is extracted}
}
\value{
the mean time at the specified milestones for 3 groups of people, the group who are 15 to 30 mins faster than your goal time, the gorup no 15 mins faster than your goal time and the group no 15 mins slower than your goal time
}
\description{
goaltime
}
\examples{
\dontrun{
goaltime(data_all)
}
}
