% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasttime.R
\name{pasttime}
\alias{pasttime}
\title{pasttime}
\usage{
pasttime(pasttime, dataset)
}
\arguments{
\item{pasttime}{the time you took to reach the finishing line in a previous marathon race}

\item{dataset}{the dataset to be treated}
}
\value{
the average time at the specified milestones based on the group of people who finished within the 15 mins range centered around your past time
}
\description{
pasttime
}
\examples{
\dontrun{
pasttime(data_all)
}
}
