% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top10percent.R
\name{top10percent}
\alias{top10percent}
\title{top10percent}
\usage{
top10percent(dataset)
}
\arguments{
\item{dataset}{the dataset to be treated, the top 10 percent runners in terms of their finishing time will be extracted}
}
\value{
the top 10 percent runners in terms of their finishing time
}
\description{
top10percent
}
\examples{
\dontrun{
top10percent(data_all)
}
}
