% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top10percentmean.R
\name{top10percentmean}
\alias{top10percentmean}
\title{top10percentmean}
\usage{
top10percentmean(dataset)
}
\arguments{
\item{dataset}{the data set to be transformed}
}
\value{
a data frame with the mean time at the specified milestones of the marathon for the top 10 runners in the dataset
}
\description{
top10percentmean
}
\examples{
\dontrun{
top10percent(data_all)
}
}
